webkit-efl/LayoutTests/fast/dom/form-data-exceptions.html

<!DOCTYPE html>
<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<script>
var formData = new FormData();
shouldThrow('formData.append()');
shouldThrow('formData.append("foo")');
</script>
</body>
<script src="../../fast/js/resources/js-test-post.js"></script>
</html>
