webkit-efl/LayoutTests/fast/dom/form-data-exceptions-expected.txt

PASS formData.append() threw exception TypeError: Not enough arguments.
PASS formData.append("foo") threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
